+++ src/main/java/org/jabref/gui/desktop/JabRefDesktop.java
@@ -1,5 +1,7 @@
 package org.jabref.gui.desktop;
 
+import java.awt.Desktop;
+import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.nio.file.Files;
@@ -306,5 +308,27 @@ public class JabRefDesktop {
             dialogService.notify(couldNotOpenBrowser);
             dialogService.showErrorDialogAndWait(couldNotOpenBrowser, couldNotOpenBrowser + "\n" + openManually + "\n" + copiedToClipboard);
         }
+    }
+
+    /**
+     * Moves the given file to the trash.
+     *
+     * @throws UnsupportedOperationException if the current platform does not support the {@link Desktop.Action#MOVE_TO_TRASH} action
+     * @see Desktop#moveToTrash(File)
+     */
+    public static void moveToTrash(Path path) {
+        NATIVE_DESKTOP.moveToTrash(path);
+    }
+
+    public static boolean moveToTrashSupported() {
+        return NATIVE_DESKTOP.moveToTrashSupported();
+    }
+
+    public static Path getApplicationDirectory() {
+        return NATIVE_DESKTOP.getApplicationDirectory();
+    }
+
+    public static Path getFulltextIndexBaseDirectory() {
+        return NATIVE_DESKTOP.getFulltextIndexBaseDirectory();
     }
 }