+++ src/main/java/org/jabref/gui/preferences/linkedfiles/LinkedFilesTabViewModel.java
@@ -40,6 +40,7 @@ public class LinkedFilesTabViewModel implements PreferenceTabViewModel {
     private final StringProperty fileNamePatternProperty = new SimpleStringProperty();
     private final StringProperty fileDirectoryPatternProperty = new SimpleStringProperty();
     private final BooleanProperty confirmLinkedFileDeleteProperty = new SimpleBooleanProperty();
+    private final BooleanProperty moveToTrashProperty = new SimpleBooleanProperty();
 
     private final Validator mainFileDirValidator;
 
@@ -82,6 +83,7 @@ public class LinkedFilesTabViewModel implements PreferenceTabViewModel {
         fileNamePatternProperty.setValue(filePreferences.getFileNamePattern());
         fileDirectoryPatternProperty.setValue(filePreferences.getFileDirectoryPattern());
         confirmLinkedFileDeleteProperty.setValue(filePreferences.confirmDeleteLinkedFile());
+        moveToTrashProperty.setValue(filePreferences.moveToTrash());
 
         // Autolink preferences
         switch (autoLinkPreferences.getCitationKeyDependency()) {
@@ -113,6 +115,7 @@ public class LinkedFilesTabViewModel implements PreferenceTabViewModel {
 
         autoLinkPreferences.setRegularExpression(autolinkRegexKeyProperty.getValue());
         filePreferences.confirmDeleteLinkedFile(confirmLinkedFileDeleteProperty.getValue());
+        filePreferences.moveToTrash(moveToTrashProperty.getValue());
     }
 
     ValidationStatus mainFileDirValidationStatus() {
@@ -184,6 +187,10 @@ public class LinkedFilesTabViewModel implements PreferenceTabViewModel {
 
     public BooleanProperty confirmLinkedFileDeleteProperty() {
         return this.confirmLinkedFileDeleteProperty;
+    }
+
+    public BooleanProperty moveToTrashProperty() {
+        return this.moveToTrashProperty;
     }
 }